checks if csv exists and creates/appends as appropriate

+++ arrests.py
@@ -2,6 +2,7 @@ import requests
 from bs4 import BeautifulSoup
 import pandas as pd
 import datetime
+import os.path
 
 # set pandas to show all columns
 pd.set_option('display.expand_frame_repr', False)
@@ -19,7 +20,7 @@ df = df[0]
 # clean up arrest time
 df['arrest'] = pd.to_datetime(df['ADMITTED_DATE'])
 
-# clean up control number (maybe just drop this?)
+# clean up control number (maybe just drop this? Not sure what it is for)
 df['control'] = df['CONTROL_NUMBER'].fillna(0).astype(int)
 
 # clean up names and split into first and last
@@ -36,20 +37,26 @@ df.rename(columns={'SEX': 'sex', 'RACE_DESC': 'race'}, inplace=True)
 # drop the columns we don't need
 df = df.drop(columns=['Details', 'RELEASE_DATE', 'ADMITTED_DATE', 'CONTROL_NUMBER', 'DATE_OF_BIRTH', 'Inmate Name'])
 
-# subset yesterday's date (so when we append, we won't duplicate. We will only add one date at a time.
+# subset yesterday's date (so when we append, we won't duplicate. We will only add one date at a time.)
 df = df[(df['arrest'] > pd.Timestamp(datetime.datetime.today().date()) - datetime.timedelta(1)) & (df['arrest'] < pd.Timestamp(datetime.datetime.today().date()))]
 
-# sort by time
+# sort df by time so it's organized and pretty
 df.sort_values(by=['arrest'], inplace=True)
 
+# reset the index so that it's clean looking. Not strictly necessary, but it will make the output csv cleaner
+# this will reset the index to whatever the last row number is in the existing csv file
 df.reset_index(inplace=True)
 df.index += pd.read_csv("./arrests.csv")['Unnamed: 0'].max() + 1
 
+# drops the index column which gets generated from the OLD index numbers when the index is reset above
 df = df.drop(columns=['index'])
 
-print(df)
+# check if the csv file already exists. If not, create a new one. If it exists, append to the existing one.
+if os.path.isfile('./arrests.csv') is True:
+    print('file exists, appending...')
+    with open('arrests.csv', 'a') as f:
+        df.to_csv(f, header=False)
 
-#df.to_csv('./arrests.csv')
-
-with open('arrests.csv', 'a') as f:
-    df.to_csv(f, header=False)
+else:
+    print('file does not exists, creating...')
+    df.to_csv('./arrests.csv')
